feat: sistema de sanciones por dias de demora + 3 strikes = 1 mes de suspension

+++ src/com/bibliotech/model/Prestamo.java
@@ -3,5 +3,5 @@ package com.bibliotech.model;
 import java.time.LocalDate;
 import java.util.Optional;
 
-public record Prestamo(int id, Libro libro, Socio socio, LocalDate fechaPrestamo, LocalDate fechaDevolucionEsperada, Optional<LocalDate> fechaDevolucionReal) {
+public record Prestamo(long id, Libro libro, Socio socio, LocalDate fechaPrestamo, LocalDate fechaDevolucionEsperada, Optional<LocalDate> fechaDevolucionReal) {
 }
